requirements.txt
flask==2.3.3
flask-cors==3.0.10
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1

psycopg2-binary==2.9.9
python-dotenv==1.0.1

torch==2.2.2
transformers==4.40.1
protobuf==3.20.3

google-generativeai==0.4.1
